MLProject/requirements.txt
absl-py==0.9.0
astor==0.8.1
Django==3.0.6
django-crispy-forms==1.9.0
djangorestframework==3.11.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.2
pandas==1.0.1
protobuf==3.11.3
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
SciPy==1.4.1
scikit-learn
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tensorboard==2.2.1
tensorflow==1.15.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.12.1
xlrd==1.2.0
